requirements.txt
selenium
pytest
pytest-html
pytest-rerunfailures
pytest-xdist
Appium-Python-Client